backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.43.0",
    "fastapi>=0.136.1",
    "openai>=2.33.0",
    "openai-agents[litellm]>=0.14.8",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
